# src/main/java/com/company/project/dto/TodoRequest.java
package com.company.project.dto;

import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.Size;

public record TodoRequest(

        @NotBlank(message = "Title is required")
        @Size(max = 255, message = "Title must be less than 255 characters")
        String title,

        @Size(max = 1000, message = "Description must be less than 1000 characters")
        String description
) {
}