# src/main/resources/api/openapi-spec.yaml
openapi: 3.0.3
info:
  title: Generic Spring Boot App API
  description: |
    Enterprise Spring Boot 3 Backend API
    
    ## Features
    - Todo Management
    - Health Check Endpoints
    - JWT Authentication (planned)
    - OAuth2 Support (planned)
    
    ## Response Format
    All endpoints return standardized responses:
    ```json
    {
      "timestamp": "2026-02-20T12:00:00Z",
      "status": 200,
      "data": {},
      "error": null
    }
    ```
  version: 1.0.0
  contact:
    name: API Support
  license:
    name: MIT License

servers:
  - url: http://localhost:8080
    description: Development server
  - url: https://api.example.com
    description: Production server

tags:
  - name: Health
    description: Health check and system endpoints
  - name: Todo
    description: Todo management operations

paths:
  /api/health:
    get:
      tags:
        - Health
      summary: Health Check
      description: Check application health status
      operationId: healthCheck
      responses:
        '200':
          description: Application is healthy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'

  /hello:
    get:
      tags:
        - Health
      summary: Hello World
      description: Simple hello endpoint
      operationId: hello
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Hello World

  /api/v1/sql/todo:
    post:
      tags:
        - Todo
      summary: Create Todo
      description: Create a new todo item
      operationId: createTodo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TodoRequest'
      responses:
        '201':
          description: Todo created successfully
          headers:
            Location:
              description: URL of the created resource
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoResponse'
        '400':
          description: Invalid input - title is required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      tags:
        - Todo
      summary: Get All Todos
      description: Retrieve a list of all todo items
      operationId: getAllTodos
      responses:
        '200':
          description: Successfully retrieved list of todos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TodoResponse'

  /api/v1/sql/todo/{id}:
    get:
      tags:
        - Todo
      summary: Get Todo by ID
      description: Retrieve a specific todo item by its ID
      operationId: getTodoById
      parameters:
        - name: id
          in: path
          required: true
          description: The unique ID of the todo
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Todo found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoResponse'
        '404':
          description: Todo not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - Todo
      summary: Update Todo
      description: Update an existing todo item with new title and description
      operationId: updateTodo
      parameters:
        - name: id
          in: path
          required: true
          description: The unique ID of the todo
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TodoRequest'
      responses:
        '200':
          description: Todo updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Todo not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Todo
      summary: Delete Todo
      description: Delete a todo item by its ID
      operationId: deleteTodo
      parameters:
        - name: id
          in: path
          required: true
          description: The unique ID of the todo
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Todo deleted successfully
        '404':
          description: Todo not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/sql/todo/{id}/toggle:
    patch:
      tags:
        - Todo
      summary: Toggle Todo Completion
      description: Toggle the completion status of a todo item
      operationId: toggleTodoCompletion
      parameters:
        - name: id
          in: path
          required: true
          description: The unique ID of the todo
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Todo completion status toggled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoResponse'
        '404':
          description: Todo not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    TodoRequest:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          description: Todo title
          maxLength: 255
          example: Sample Todo
        description:
          type: string
          description: Todo description
          maxLength: 1000
          example: This is a sample todo item

    TodoResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Unique identifier
          example: 1
        title:
          type: string
          description: Todo title
          example: Sample Todo
        description:
          type: string
          description: Todo description
          example: This is a sample todo item
        completed:
          type: boolean
          description: Completion status
          example: false
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp
          example: "2026-02-20T12:00:00Z"
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
          example: "2026-02-20T12:00:00Z"

    HealthResponse:
      type: object
      properties:
        status:
          type: string
          description: Health status
          example: UP
        timestamp:
          type: string
          format: date-time
          description: Response timestamp
          example: "2026-02-20T12:00:00Z"

    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
        status:
          type: integer
          description: HTTP status code
        error:
          type: string
          description: Error message
        path:
          type: string
          description: Request path

  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer token authentication
    oauth2:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://api.example.com/oauth/authorize
          tokenUrl: https://api.example.com/oauth/token
          scopes:
            read: Read access
            write: Write access

security:
  - bearerAuth: []
  - oauth2:
      - read
      - write